+++ lib/ui/views/home/home_view.dart
@@ -6,6 +6,7 @@ import 'package:stacked/stacked.dart';
 import 'package:veggie_go_malaysia/constants/colors.dart';
 import 'package:veggie_go_malaysia/ui/views/home/home_viewmodel.dart';
 import 'package:veggie_go_malaysia/ui/views/home/widgets/location_bar.dart';
+import 'package:veggie_go_malaysia/ui/views/home/widgets/quick_search.dart';
 
 class HomeView extends StatelessWidget {
   @override
@@ -23,7 +24,7 @@ class HomeView extends StatelessWidget {
                   child: _SearchBar(),
                 ),
                 _AnnouncementCarousel(),
-                _QuickSearch(),
+                QuickSearch(),
                 _FilterResults(),
                 _ResultsListView(),
               ],
@@ -66,57 +67,6 @@ class _AnnouncementCarousel extends StatelessWidget {
         autoPlay: true,
         enableInfiniteScroll: false,
         height: 500.h,
-      ),
-    );
-  }
-}
-
-class _QuickSearch extends StatelessWidget {
-  static Widget _customIcon(Color color, String text) {
-    return SizedBox(
-      width: 220.w,
-      child: Column(
-        children: <Widget>[
-          CircleAvatar(
-            radius: 80.w,
-            backgroundColor: color,
-          ),
-          SizedBox(height: 10.h),
-          Text(text),
-        ],
-      ),
-    );
-  }
-
-  final List<Widget> _quickIconsFirst = [
-    _customIcon(ThemeColors.searchBarIcon, 'Japanese'),
-    _customIcon(ThemeColors.searchBarIcon, 'Chinese'),
-    _customIcon(ThemeColors.searchBarIcon, 'Western'),
-    _customIcon(ThemeColors.searchBarIcon, 'Indian'),
-  ];
-  final List<Widget> _quickIconsSecond = [
-    _customIcon(ThemeColors.searchBarIcon, 'Korean'),
-    _customIcon(ThemeColors.searchBarIcon, 'Thai'),
-    _customIcon(ThemeColors.searchBarIcon, 'Mixed'),
-    _customIcon(ThemeColors.searchBarIcon, 'More'),
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.symmetric(vertical: 50.h),
-      child: Column(
-        children: <Widget>[
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: _quickIconsFirst,
-          ),
-          SizedBox(height: 40.h),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: _quickIconsSecond,
-          )
-        ],
       ),
     );
   }
